+++ src/mito2/src/test_util.rs
@@ -24,6 +24,7 @@ use log_store::raft_engine::log_store::RaftEngineLogStore;
 use log_store::test_util::log_store_util;
 use object_store::services::Fs;
 use object_store::ObjectStore;
+use store_api::metadata::{ColumnMetadata, SemanticType};
 use store_api::storage::RegionId;
 
 use crate::config::MitoConfig;
@@ -31,7 +32,7 @@ use crate::engine::MitoEngine;
 use crate::error::Result;
 use crate::manifest::manager::RegionManifestManager;
 use crate::manifest::options::RegionManifestOptions;
-use crate::metadata::{ColumnMetadata, RegionMetadataRef, SemanticType};
+use crate::metadata::RegionMetadataRef;
 use crate::request::{CreateRequest, RegionOptions};
 use crate::worker::WorkerGroup;
